pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Personal AI Employee — Bronze Tier (Hackathon 0)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Silver tier: Gmail + WhatsApp watchers
silver = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "playwright>=1.40.0",
]

[project.scripts]
fs-watcher   = "watchers.filesystem_watcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["watchers"]
